Edit del obj funcs

+++ db_work.py
@@ -165,7 +165,7 @@ class db_work():
 
 		Повертає [col_id]
 		"""
-		
+
 		res = self.__cur.execute(f'''SELECT cols_order
 									FROM v_{element}_cols
 									WHERE {element}_id = {element_id}''').fetchone()
@@ -260,23 +260,27 @@ class db_work():
 		# перебір подій
 		command_events = self.__cur.execute(f'''SELECT event_id FROM commands_event
 									WHERE command_id = {command_id}''').fetchall()
-		for event in command_events:
-			event = event['event_id']
-			self.__cur.execute(f'DELETE FROM events WHERE event_id = {event}')
+		if command_events:
+			for event in command_events:
+				event = event['event_id']
+				self.__cur.execute(f'DELETE FROM events WHERE event_id = {event}')
 
 		# перебір колонок
 		command_cols = self.__cur.execute(f'''SELECT cols_order FROM v_command_cols
 									WHERE command_id = {command_id}''').fetchall()
-		for col in command_cols:
-			col = col['cols_order'].split(',')
+		if command_cols:
+			for col in command_cols:
+				col = col['cols_order'].split(',')
 
-			# перебір завдань в колонках
-			for col_id in col:
-				command_tasks = self.__cur.execute(f'''SELECT task_id FROM commands_task
-									WHERE command_id = {command_id}''').fetchall()
-				for task in command_tasks:
-					task = task['task_id']
-					self.__cur.execute(f'DELETE FROM tasks WHERE task_id = {task}')
+				# перебір завдань в колонках
+				for col_id in col:
+					command_tasks = self.__cur.execute(f'''SELECT task_id FROM commands_task
+										WHERE command_id = {command_id}''').fetchall()
+					if command_tasks:
+						for task in command_tasks:
+							task = task['task_id']
+							self.__cur.execute(f'DELETE FROM tasks WHERE task_id = {task}')
+
 		self.__cur.execute(f'DELETE FROM commands WHERE command_id = {command_id}')
 
 
@@ -300,14 +304,16 @@ class db_work():
 		if group_cols:
 			for col in group_cols:
 				col = col['cols_order'].split(',')
+
 				# перебір завдань колонок
 				for col_id in col:
 					group_tasks = self.__cur.execute(f'''SELECT task_id FROM groups_task
 									WHERE group_id = {group_id}''').fetchall()
-							
-					for task in group_tasks:
-						task = task['task_id']
-						self.__cur.execute(f'DELETE FROM tasks WHERE task_id = {task}')
+					
+					if group_tasks:
+						for task in group_tasks:
+							task = task['task_id']
+							self.__cur.execute(f'DELETE FROM tasks WHERE task_id = {task}')
 							
 					self.__cur.execute(f'DELETE FROM cols WHERE col_id = {col_id}')
 		self.__cur.execute(f'DELETE FROM groups WHERE group_id = {group_id}')
